[GT-1857] Fix iOS app crash issue caused by Google Signin
This commit fixes a blocker issue when user attempt to
authenticate with an account without complete user information
fields (e.g empty firstname, name, or family name).

Refer to the solution [on Github](https://github.com/react-native-community/react-native-google-signin/pull/505)

As the latest version of per library is now `1.00rc` which might
have several breaking changes which it's risky to upgrade the whole
library instead of simply change this small piece of code.